Fix mobignore status behaving backwards
The original code from 11 years ago for the /mobignore command was
incorrect in its behaviour, as it would set the configuration value
to the opposite of the current value, then print the old config value.

More recently, the event handler for mob ignore was incorrectly inverted
to match the confusing behaviour, effectively changing the default
behaviour and causing all players to be ignored by mobs. In survival.