Cache data for /backs to improve performance
Previously, every time a player ran /backs, it would take time for the
server to load each location and determine information about it (eg.
the block the player was standing on) which would visually lag in the
chat, and could introduce minor lag to all other players on the server -
as it was loading multiple chunks at the same time, likely far apart.
Instead, do this during the initial creation of the item in the back
stack, while the location is still in memory.