Fixed: Attribute list and delete commands
Loop through attributes and then their modifiers, so that a reference
to the original attribute is available for displaying the name and for
being able to delete.

Rather than generating a full UUID, use half of one.

This also fixes the tooltip formatting and the display of the item name.

Fixes #35

The remove command still needs to be adjusted to either allow deletion
of all modifiers for each attribute, or individual modifiers.