Updated: Directory validation for /wm create and load
This fixes a number of quirks, and provides better error messages.

Previously, loading a world happened blindly, so if a directory existed
but wasn't a world, it'd implicitly initialise as one, which is not ideal.
The same would happen when creating a world, with the additional bonus
of the create flow loading valid world directories when it shouldn't.

This also adds valid world directories to tab completion of /wm load