Updated: Restore CurseForge Update Checker alongside Spigot
This deprecates custom_tags in favour of platform-specific tags.
If you are currently using custom_tags, they will continue to work
forever, with the following caveat:
- if the tag is a string, it'll try BukkitDev/CurseForge via your API key
- if the tag is a number, it'll check SpigotMC

The BukkitDev site, while no longer providing the old RSS feed, is
powered by CurseForge. This means there's a functional JSON API, which
we can use to find plugins and check the latest release.

Code has been split up for readability, and fetches the following:
1. SpigotMC via a numeric ID, which is preferred.
     This is the final number in the URL when browsing Resources.
2. CurseForge/BukkitDev via both string (legacy) and ID
     The ID can be found on the CurseForge page under "Project ID"
     The string can be determined by the URL on either site, as always

Handling of JSON exceptions is relegated to the top level, so that the
user is informed the check failed in some way. Better error bubbling
could be a good future enhancement.

There are no plans to bring back the automatic downloader, as it'd be
too much to maintain and not fantastic for security.

Additional plugin hosting services could be added in future.