Fixed: /whitelist remove players not in memory
Annoyingly, if you getOfflinePlayer(uuid) it doesn't lookup the name,
but getOfflinePlayer(name) does and stores it in memory. So if an unknown
player is added to the whitelist, and the remove command is run in the
same session, everything works fine. But if you restart the server in
between, everything goes bad.

This updates the tab completion to show UUIDs for unknown players, and
for the Removed message to display the UUID instead of erroring out.