Updated: PacketEvents to replace ProtocolLib
This makes the packet handling code easier to maintain.

This also fixes an issue where inventory items would be invisible, which
was caused by WINDOW_ITEMS setting null or continuing when it should
have set the unmodified item stack back to the slot.

This whole system is mostly kept for compatibility with old configs.